backend/requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
python-dotenv==1.0.1
pymongo==4.6.1
dnspython==2.4.2
flask-cors==4.0.0 